performance-timeline/tentative/include-frames-two-local-children.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
</body>
<script>
promise_test(() => {
    return new Promise(resolve => {
        performance.clearResourceTimings()

        // Create first child iframe.
        const childFrameOne = document.createElement('iframe')
        childFrameOne.src = "../resources/child-frame.html"
        document.body.appendChild(childFrameOne)

        childFrameOne.addEventListener("load", () => {
            window.top.postMessage("DONE", "*")
        })

        // Create second child iframe.
        const childFrameTwo = document.createElement('iframe')
        childFrameTwo.src = "../resources/child-frame.html"
        document.body.appendChild(childFrameTwo)

        childFrameTwo.addEventListener("load", () => {
            window.top.postMessage("DONE", "*")
        })

        // Call getEntries() once the parent frame recieves two messages.
        let messageCount = 0
        window.addEventListener("message", (e) => {
            messageCount++
            if (messageCount > 1) {
                const entries = performance.getEntries(true)
                const navigationEntries = performance.getEntriesByType("navigation", true)
                const markedEntries = performance.getEntriesByName("entry-name", undefined, true)

                // 4 entries for parent, 2 for each child
                assert_equals(entries.length, 8)

                // 1 entry for parent, 1 for each child.
                assert_equals(navigationEntries.length, 3)

                // 1 entry for each child.
                assert_equals(markedEntries.length, 2)

                resolve()
            }
        })
    })
}, "GetEntries of a Parent Frame with two LocalFrame children")
</script>
